Fix table deletion bug in run_tests_with_args function

+++ tests/literal_utils.py
@@ -120,7 +120,8 @@ def run_tests_with_args(
         # Adjust to the structure of LiteralTestValues
 
         table_as_dict = test.vals_for_df
-        db_api._delete_table_from_database("__splink__test_table")
+        if db_api.table_exists_in_database("__splink__test_table"):
+            db_api._delete_table_from_database("__splink__test_table")
         db_api._table_registration(table_as_dict, "__splink__test_table")
 
         if test.expected_exception:
